Moved EspooIncomeCoefficientMultiplierProvider under espoo specific folder.
